linting and tests
