+++ .circleci/config.yml
@@ -683,11 +683,13 @@ workflows:
             - secrets-check
 
   dev-deployment:
+    when:
+      << pipeline.parameters.run_dev_deployment >>
     jobs:
       - deploy-infrastructure-dev:
-          target_env: sandbox
+          target_env: << pipeline.parameters.target_env >>
       - deploy-dev:
-          target_env: sandbox
+          target_env: << pipeline.parameters.target_env >>
           requires:
             - deploy-infrastructure-dev
