+++ tdrs-backend/tdpservice/scheduling/test/test_user_deactivation_warning.py
@@ -1,3 +1,4 @@
+"""Test functions for deactivated ."""
 from django.utils import timezone
 
 import pytest
@@ -13,7 +14,7 @@ logger = logging.getLogger(__name__)
 def test_deactivation_email_10_days(user, mocker):
     """Test that the check_for_accounts_needing_deactivation_warning task runs."""
     mocker.patch(
-        'tdpservice.email.email.send_deactivation_warning_email',
+        'tdpservice.email.email_helper.send_deactivation_warning_email',
         return_value=None
     )
 
@@ -21,13 +22,13 @@ def test_deactivation_email_10_days(user, mocker):
     user.first_name = 'UniqueName'
     user.save()
     check_for_accounts_needing_deactivation_warning()
-    assert tdpservice.email.email.send_deactivation_warning_email.called_once_with(users = [user], days = 10)
+    assert tdpservice.email.email_helper.send_deactivation_warning_email.called_once_with(users=[user], days=10)
 
 @pytest.mark.django_db
 def test_deactivation_email_3_days(user, mocker):
     """Test that the check_for_accounts_needing_deactivation_warning task runs."""
     mocker.patch(
-        'tdpservice.email.email.send_deactivation_warning_email',
+        'tdpservice.email.email_helper.send_deactivation_warning_email',
         return_value=None
     )
 
@@ -35,13 +36,13 @@ def test_deactivation_email_3_days(user, mocker):
     user.first_name = 'UniqueName'
     user.save()
     check_for_accounts_needing_deactivation_warning()
-    assert tdpservice.email.email.send_deactivation_warning_email.called_once_with(users = [user], days = 3)
+    assert tdpservice.email.email_helper.send_deactivation_warning_email.called_once_with(users=[user], days=3)
 
 @pytest.mark.django_db
 def test_deactivation_email_1_days(user, mocker):
     """Test that the check_for_accounts_needing_deactivation_warning task runs."""
     mocker.patch(
-        'tdpservice.email.email.send_deactivation_warning_email',
+        'tdpservice.email.email_helper.send_deactivation_warning_email',
         return_value=None
     )
 
@@ -49,14 +50,14 @@ def test_deactivation_email_1_days(user, mocker):
     user.first_name = 'UniqueName'
     user.save()
     check_for_accounts_needing_deactivation_warning()
-    assert tdpservice.email.email.send_deactivation_warning_email.called_once_with(users = [user], days = 1)
-    
+    assert tdpservice.email.email_helper.send_deactivation_warning_email.called_once_with(users=[user], days=1)
+
 
 @pytest.mark.django_db
 def test_no_users_to_warn(user, mocker):
     """Test that the check_for_accounts_needing_deactivation_warning task runs."""
     mocker.patch(
-        'tdpservice.email.email.send_deactivation_warning_email',
+        'tdpservice.email.email_helper.send_deactivation_warning_email',
         return_value=None
     )
 
@@ -64,4 +65,4 @@ def test_no_users_to_warn(user, mocker):
     user.first_name = 'UniqueName'
     user.save()
     check_for_accounts_needing_deactivation_warning()
-    tdpservice.email.email.send_deactivation_warning_email.assert_not_called()
+    tdpservice.email.email_helper.send_deactivation_warning_email.assert_not_called()
